requirements.txt
# ============================================================
# FinSight — Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# --- Agent Framework ---
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0

# --- Financial Data ---
yfinance>=0.2.40
requests>=2.31.0

# --- Data Processing & Analysis ---
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0

# --- MLOps: Experiment Tracking ---
mlflow>=2.13.0

# --- MLOps: Monitoring & Drift Detection ---
evidently>=0.4.30

# --- PDF Report Generation ---
reportlab>=4.0.0

# --- Frontend ---
streamlit>=1.35.0

# --- Configuration ---
python-dotenv>=1.0.0

# --- Utilities ---
tqdm>=4.66.0

# --- Testing ---
pytest>=8.0.0
pytest-cov>=5.0.0
